+++ athetes_unlimited_py/softball.py
@@ -28,7 +28,7 @@ def get_au_softball_season(season_id:int) -> int:
     ----------
     The proper season corresponding to an Athletes Unlimited softball season ID.
     """
-    season = 0
+    #season = 0
     
     if season_id == 2:
         season= 2020
@@ -49,7 +49,7 @@ def get_au_softball_season(season_id:int) -> int:
         season = 2023
         return season 
     else:
-        raise ValueError(f'[season] can only be 2022 or 2023 at this time for softball.\nYou entered :\n\t{season}')
+        raise ValueError(f'[season] can only be 2022 or 2023 at this time for softball.\nYou entered :\n\t{season_id}')
 
 def get_au_softball_season_id(season:int) -> int:
     """
@@ -178,6 +178,7 @@ def get_au_softball_game_stats(season_id:int,game_num:int,get_team_stats=False,g
             row_df['season_type'] = i['battingStats'][0]['gamesStarted']
             row_df['G'] = i['battingStats'][0]['gamesPlayed']
             row_df['GS'] = i['battingStats'][0]['gamesStarted']
+            row_df['batting_PA'] = 0
             row_df['batting_AB'] = i['battingStats'][0]['atBat']
             row_df['batting_R'] = i['battingStats'][0]['runs']
             row_df['batting_H'] = i['battingStats'][0]['hits']
@@ -198,6 +199,8 @@ def get_au_softball_game_stats(season_id:int,game_num:int,get_team_stats=False,g
             row_df['batting_SF'] = i['battingStats'][0]['sacrificeFly']
             row_df['batting_SH'] = i['battingStats'][0]['sacrificeHit']
             row_df['AU_POINTS'] = i['battingStats'][0]['auTotalPoints']
+
+            row_df['batting_PA'] = row_df['batting_AB'] + row_df['batting_BB'] + row_df['batting_HBP'] + row_df['batting_SF'] + row_df['batting_SH']
         else:
             row_df['batting_AB'] = None
             row_df['batting_R'] = None
@@ -236,12 +239,14 @@ def get_au_softball_game_stats(season_id:int,game_num:int,get_team_stats=False,g
             row_df['pitching_CG'] = i['pitchingStats'][0]['completeGames']
             row_df['pitching_SV'] = i['pitchingStats'][0]['saves']
             row_df['pitching_IP_str'] = str(i['pitchingStats'][0]['inningsPitched'])
+            row_df['pitching_IP_str'] = row_df['pitching_IP_str'].replace('None',None)
 
             try:
                 row_df['pitching_IP'] = float(str(i['pitchingStats'][0]['inningsPitched']).replace('.1','.333').replace('.2','.667'))
             except:
                 row_df['pitching_IP'] = None
 
+            row_df['pitching_QS'] = 0
             row_df['pitching_H'] = i['pitchingStats'][0]['hits']
             row_df['pitching_R'] = i['pitchingStats'][0]['runs']
             row_df['pitching_ER'] = i['pitchingStats'][0]['earnedRuns']
@@ -255,11 +260,12 @@ def get_au_softball_game_stats(season_id:int,game_num:int,get_team_stats=False,g
             row_df['pitching_HR9'] = (9 * row_df['pitching_HR']) / row_df['pitching_IP']
             row_df['pitching_BB9'] = (9 * row_df['pitching_BB']) / row_df['pitching_IP']
             row_df['pitching_SO9'] = (9 * row_df['pitching_SO']) / row_df['pitching_IP']
-            row_df['pitching_SO/BB'] = (9 * row_df['pitching_SO']) / row_df['pitching_BB']
+            row_df['pitching_SO/BB'] = row_df['pitching_SO'] / row_df['pitching_BB']
             row_df['pitching_RA9'] = 9 * (row_df['pitching_R'] / row_df['pitching_IP'])
             row_df['pitching_PI'] = i['pitchingStats'][0]['numberOfPitches']
             row_df['pitching_PI_balls'] = i['pitchingStats'][0]['balls']
             row_df['pitching_PI_strikes'] = i['pitchingStats'][0]['strikes']
+            row_df['pitcing_game_score'] = 50 + (row_df['pitching_IP'] * 3) + row_df['pitching_SO'] - (row_df['pitching_H'] * 2) - (row_df['pitching_ER'] * 4) - ((row_df['pitching_R'] - row_df['pitching_ER']) * 2) - row_df['pitching_BB']
             row_df['AU_POINTS'] = i['pitchingStats'][0]['auTotalPoints']
         else:
             row_df['pitching_H'] = None
@@ -329,10 +335,6 @@ def get_au_softball_game_stats(season_id:int,game_num:int,get_team_stats=False,g
         ## Save the data to the correct DataFrame
         ##############################################################################################################################
 
-        if i['type'] == "Team":
-            team_stats_df = pd.concat([team_stats_df,row_df],ignore_index=True)
-        else:
-            player_stats_df = pd.concat([player_stats_df,row_df],ignore_index=True)
 
         row_df['type'] = i['type']
         row_df['teamId'] = i['teamId']
@@ -342,7 +344,14 @@ def get_au_softball_game_stats(season_id:int,game_num:int,get_team_stats=False,g
         else:
             row_df['homeTeamFlg'] = 0
 
+        if i['type'] == "Team":
+            team_stats_df = pd.concat([team_stats_df,row_df],ignore_index=True)
+        else:
+            player_stats_df = pd.concat([player_stats_df,row_df],ignore_index=True)
+
         del row_df
+
+    player_stats_df.loc[(player_stats_df['pitching_IP'] >= 6) & (player_stats_df['GS'] == 1) & (player_stats_df['pitching_ER'] <= 3),'pitching_QS'] = 1
 
     ##############################################################################################################################
     ## Once we're done, return the correct dataframe.
@@ -512,7 +521,7 @@ def get_au_softball_season_pbp(season:int) -> pd.DataFrame():
         if i['seasonId'] == seasonId:
             len_game_ids = len(i['gameIds'])
 
-            for j in tqdm(range(1,len_game_ids+1)):
+            for j in tqdm(i['gameIds']):
                 print(f'\nOn game {j} of {len_game_ids+1} for {season}.')
                 # try:
                 #     game_df = get_softball_game_stats(season,j)
@@ -520,7 +529,7 @@ def get_au_softball_season_pbp(season:int) -> pd.DataFrame():
                 #     print(f'Couldn\'t parse game stats for game #{j}.')
                 #     time.sleep(10)
 
-                game_df = get_au_softball_game_stats(season,j,get_team_stats=True)
+                game_df = get_au_softball_pbp(seasonId,j)
 
                 season_pbp_df = pd.concat([season_pbp_df,game_df],ignore_index=True)
                 del game_df
@@ -553,17 +562,17 @@ def get_au_softball_season_player_box(season:int) -> pd.DataFrame():
     for i in sport_json_data['data']:
         #print(i)
         if i['seasonId'] == seasonId:
-            #len_game_ids = len(i['gameIds'])
+            len_game_ids = len(i['gameIds'])
 
-            for j in tqdm(i['gameIds']):
-                print(f'\nOn game ID {j} for {season}.')
+            for j in tqdm(range(1,len_game_ids+1)):
+                print(f'\nOn game {j} of {len_game_ids+1} for {season}.')
                 # try:
                 #     game_df = get_softball_game_stats(season,j)
                 # except:
                 #     print(f'Couldn\'t parse game stats for game #{j}.')
                 #     time.sleep(10)
 
-                game_df = get_au_softball_pbp(season,j)
+                game_df = get_au_softball_game_stats(seasonId,j,get_team_stats=False)
 
                 season_stats_df = pd.concat([season_stats_df,game_df],ignore_index=True)
                 del game_df
@@ -606,10 +615,278 @@ def get_au_softball_season_team_box(season:int) -> pd.DataFrame():
                 #     print(f'Couldn\'t parse game stats for game #{j}.')
                 #     time.sleep(10)
 
-                game_df = get_au_softball_game_stats(season,j,get_team_stats=True)
+                game_df = get_au_softball_game_stats(seasonId,j,get_team_stats=True)
 
                 season_stats_df = pd.concat([season_stats_df,game_df],ignore_index=True)
                 del game_df
 
     return season_stats_df
 
+############################################################################################################################################################################################################################################################
+##
+## Season Stats
+##
+############################################################################################################################################################################################################################################################
+
+def get_au_softball_season_player_stats(season:int) -> pd.DataFrame():
+    """
+    Given an Atheltes Unlimited (AU) softball season, get all season player stats for an AU softball season.
+
+    Parameters
+    ----------
+    `season` (int, mandatory):
+        The AU softball season you want season player stats from.
+
+    Returns
+    ----------
+    A pandas DataFrame containing season player stats a AU season.
+
+    """
+    game_stats_df = get_au_softball_season_player_box(season)
+    # ['sport', 'api_version', 'season', 'seasonId', 'weekNumber', 
+    #    'gameNumber', 'seasonType', 'playerId', 'uniformNumber',    
+    #    'uniformNumberDisplay', 'primaryPositionLk', 'secondaryPositionLk',
+    #    'first_name', 'last_name', 'full_name', 'week', 'game_num', 
+    #    'season_type', 'G', 'GS', 'batting_PA', 'batting_AB', 'batting_R',
+    #    'batting_H', 'batting_2B', 'batting_3B', 'batting_HR', 'batting_RBI',
+    #    'batting_BB', 'batting_HBP', 'batting_K', 'batting_SB', 'batting_SBA',
+    #    'batting_CS', 'batting_BA', 'batting_OBP', 'batting_SLG', 'batting_TB',
+    #    'batting_SF', 'batting_SH', 'AU_POINTS', 'pitching_W', 'pitching_L',
+    #    'pitching_ERA', 'pitching_SHO', 'pitching_CG', 'pitching_SV',
+    #    'pitching_IP_str', 'pitching_IP', 'pitching_H', 'pitching_R',
+    #    'pitching_ER', 'pitching_HR', 'pitching_BB', 'pitching_SO',
+    #    'pitching_HBP', 'pitching_WP', 'pitching_WHIP', 'pitching_H9',
+    #    'pitching_HR9', 'pitching_BB9', 'pitching_SO9', 'pitching_SO/BB',
+    #    'pitching_RA9', 'pitching_PI', 'pitching_PI_balls',
+    #    'pitching_PI_strikes', 'fielding_position', 'fielding_IP_str',
+    #    'fielding_IP', 'fielding_PO', 'fielding_A', 'fielding_E', 'fielding_DP',
+    #    'fielding_FLD%', 'fielding_CS', 'fielding_CS%', 'fielding_TC',
+    #    'fielding_CH', 'fielding_RF/9']
+
+    finished_df = game_stats_df.groupby(['sport', 'season', 'seasonId', 'playerId', 
+        'first_name', 'last_name', 'full_name'],as_index=False)[[
+        'G', 'GS', 'AU_POINTS',
+        'batting_PA', 'batting_AB', 'batting_R',
+        'batting_H', 'batting_2B', 'batting_3B', 'batting_HR', 'batting_RBI',
+        'batting_BB', 'batting_HBP', 'batting_K', 'batting_SB', 'batting_SBA',
+        'batting_CS', 'batting_TB', 'batting_SF', 'batting_SH', 
+        'pitching_W', 'pitching_L', 'pitching_SHO', 'pitching_CG', 'pitching_QS',
+        'pitching_SV', 'pitching_IP', 
+        'pitching_H', 'pitching_R', 'pitching_ER', 'pitching_HR', 'pitching_BB', 
+        'pitching_SO', 'pitching_HBP', 'pitching_WP', 'pitching_PI', 'pitching_PI_balls',
+        'pitching_PI_strikes',
+        'fielding_IP', 'fielding_PO', 'fielding_A', 'fielding_E', 'fielding_DP',
+        'fielding_CS', 'fielding_TC'
+    ]].sum()
+
+    finished_df[['G', 'GS', 'AU_POINTS','batting_PA', 'batting_AB', 'batting_R','batting_H', 'batting_2B', 'batting_3B', 'batting_HR', 'batting_RBI','batting_BB', 'batting_HBP', 'batting_K', 'batting_SB', 'batting_SBA','batting_CS', 'batting_TB', 'batting_SF', 'batting_SH', 'pitching_W', 'pitching_L', 'pitching_SHO', 'pitching_CG', 'pitching_QS','pitching_SV', 'pitching_H', 'pitching_R', 'pitching_ER', 'pitching_HR', 'pitching_BB', 'pitching_SO', 'pitching_HBP', 'pitching_WP', 'pitching_PI', 'pitching_PI_balls','pitching_PI_strikes','fielding_IP', 'fielding_PO', 'fielding_A', 'fielding_E', 'fielding_DP','fielding_CS', 'fielding_TC']] = finished_df[['G', 'GS', 'AU_POINTS','batting_PA', 'batting_AB', 'batting_R','batting_H', 'batting_2B', 'batting_3B', 'batting_HR', 'batting_RBI','batting_BB', 'batting_HBP', 'batting_K', 'batting_SB', 'batting_SBA','batting_CS', 'batting_TB', 'batting_SF', 'batting_SH', 'pitching_W', 'pitching_L', 'pitching_SHO', 'pitching_CG', 'pitching_QS','pitching_SV','pitching_H', 'pitching_R', 'pitching_ER', 'pitching_HR', 'pitching_BB', 'pitching_SO', 'pitching_HBP', 'pitching_WP', 'pitching_PI', 'pitching_PI_balls','pitching_PI_strikes','fielding_IP', 'fielding_PO', 'fielding_A', 'fielding_E', 'fielding_DP','fielding_CS', 'fielding_TC']].astype('int')
+    finished_df['pitching_IP'] = finished_df['pitching_IP'].astype('float')
+    ## Batting
+    finished_df.loc[finished_df['batting_AB'] >= 1, 'batting_BA'] = finished_df['batting_H'] / finished_df['batting_AB']
+    finished_df['batting_BA'] = finished_df['batting_BA'].round(3)
+
+    finished_df.loc[finished_df['batting_AB'] > 0,'batting_OBP'] = (finished_df['batting_H'] + finished_df['batting_BB'] + finished_df['batting_HBP']) / (finished_df['batting_AB'] + finished_df['batting_BB'] + finished_df['batting_HBP'] + finished_df['batting_SF'])
+    finished_df['batting_OBP'] = finished_df['batting_OBP'].round(3)
+
+    finished_df.loc[finished_df['batting_AB'] > 0,'batting_SLG'] = finished_df['batting_TB'] / finished_df['batting_AB']
+    finished_df['batting_SLG'] = finished_df['batting_SLG'].round(3)
+
+    finished_df.loc[finished_df['batting_AB'] > 0,'batting_OPS'] = finished_df['batting_OBP'] + finished_df['batting_SLG']
+    finished_df['batting_OPS'] = finished_df['batting_OPS'].round(3)
+
+    finished_df['batting_OPS+'] = None
+
+    finished_df.loc[finished_df['batting_AB'] > 0,'batting_SecA'] = (finished_df['batting_BB'] + (finished_df['batting_TB'] - finished_df['batting_H']) + (finished_df['batting_SB'] - finished_df['batting_CS'])) / finished_df['batting_AB']
+    finished_df['batting_SecA'] = finished_df['batting_SecA'].round(3)
+
+    finished_df.loc[finished_df['batting_PA'] > 0,'batting_BB%'] = finished_df['batting_BB'] / finished_df['batting_PA']
+    finished_df['batting_BB%'] = finished_df['batting_BB%'].round(3)
+
+    finished_df.loc[finished_df['batting_PA'] > 0,'batting_K%'] = finished_df['batting_K'] / finished_df['batting_PA']
+    finished_df['batting_K%'] = finished_df['batting_K%'].round(3)
+    
+    finished_df.loc[finished_df['batting_AB'] > 0,'batting_ISO'] = (finished_df['batting_TB'] - finished_df['batting_H']) / finished_df['batting_AB']
+    finished_df['batting_ISO'] = finished_df['batting_ISO'].round(3)
+    
+    finished_df.loc[finished_df['batting_AB'] > 0,'batting_BABIP'] = (finished_df['batting_H'] + finished_df['batting_HR']) / (finished_df['batting_AB'] - finished_df['batting_K'] - finished_df['batting_HR'] + finished_df['batting_SF'])
+    finished_df['batting_BABIP'] = finished_df['batting_BABIP'].round(3)
+
+    finished_df.loc[(finished_df['batting_SB'] > 0) | (finished_df['batting_HR'] > 0),'batting_PSN'] = (2 * finished_df['batting_HR'] * finished_df['batting_SB']) / (finished_df['batting_HR'] * finished_df['batting_SB'])
+    finished_df['batting_PSN'] = finished_df['batting_PSN'].round(3)
+
+    ## Pitching
+    finished_df['pitching_ERA'] = 9 * (finished_df['pitching_ER'] / finished_df['pitching_IP'])
+    finished_df['pitching_ERA'] = finished_df['pitching_ERA'].round(3)
+
+    finished_df['pitching_ERA+'] = None
+    finished_df['pitching_FIP'] = None
+    finished_df['pitching_FIP-'] = None
+    finished_df['pitching_WHIP'] = (finished_df['pitching_BB'] + finished_df['pitching_H']) / finished_df['pitching_IP']
+    finished_df['pitching_WHIP'] = finished_df['pitching_WHIP'].round(3)
+
+    finished_df['pitching_H9'] = (9 * finished_df['pitching_H']) / finished_df['pitching_IP']
+    finished_df['pitching_H9'] = finished_df['pitching_H9'].round(3)
+
+    finished_df['pitching_HR9'] = (9 * finished_df['pitching_HR']) / finished_df['pitching_IP']
+    finished_df['pitching_HR9'] = finished_df['pitching_HR9'].round(3)
+
+    finished_df['pitching_BB9'] = (9 * finished_df['pitching_BB']) / finished_df['pitching_IP']
+    finished_df['pitching_BB9'] = finished_df['pitching_BB9'].round(3)
+
+    finished_df['pitching_SO9'] = (9 * finished_df['pitching_SO']) / finished_df['pitching_IP']
+    finished_df['pitching_SO9'] = finished_df['pitching_SO9'].round(3)
+
+    finished_df['pitching_SO/BB'] = finished_df['pitching_SO'] / finished_df['pitching_BB']
+    finished_df['pitching_SO/BB'] = finished_df['pitching_SO/BB'].round(3)
+
+    finished_df['pitching_RA9'] = 9 * (finished_df['pitching_R'] / finished_df['pitching_IP'])
+    finished_df['pitching_RA9'] = finished_df['pitching_RA9'].round(3)
+    
+    ## Fielding
+    finished_df['fielding_FLD%'] = (finished_df['fielding_PO'] + finished_df['fielding_A']) / (finished_df['fielding_PO'] + finished_df['fielding_A'] + finished_df['fielding_E'])
+    finished_df['fielding_FLD%'] = finished_df['fielding_FLD%'].round(3)
+
+    finished_df['fielding_CH'] = finished_df['fielding_PO'] + finished_df['fielding_A'] + finished_df['fielding_E']
+    finished_df['fielding_CH'] = finished_df['fielding_CH'].round(3)
+
+    #finished_df['fielding_CS%'] = 0
+    #finished_df['fielding_CS%'] = finished_df['fielding_CS%'].round(3)
+
+    finished_df['fielding_RF/9'] = (9 * (finished_df['fielding_PO'] + finished_df['fielding_A'])) / finished_df['fielding_IP']
+    finished_df['fielding_RF/9'] = finished_df['fielding_RF/9'].round(3)
+
+    # finished_df.to_csv('test.csv')
+    # print(finished_df)
+    # print(game_stats_df.columns)
+    return finished_df
+
+def get_au_softball_season_team_stats(season:int) -> pd.DataFrame():
+    """
+    Given an Atheltes Unlimited (AU) softball season, get all season team stats for an AU softball season.
+
+    Parameters
+    ----------
+    `season` (int, mandatory):
+        The AU softball season you want season player stats from.
+
+    Returns
+    ----------
+    A pandas DataFrame containing season player stats a AU season.
+
+    """
+    game_stats_df = get_au_softball_season_player_box(season)
+    # ['sport', 'api_version', 'season', 'seasonId', 'weekNumber', 
+    #    'gameNumber', 'seasonType', 'playerId', 'uniformNumber',    
+    #    'uniformNumberDisplay', 'primaryPositionLk', 'secondaryPositionLk',
+    #    'first_name', 'last_name', 'full_name', 'week', 'game_num', 
+    #    'season_type', 'G', 'GS', 'batting_PA', 'batting_AB', 'batting_R',
+    #    'batting_H', 'batting_2B', 'batting_3B', 'batting_HR', 'batting_RBI',
+    #    'batting_BB', 'batting_HBP', 'batting_K', 'batting_SB', 'batting_SBA',
+    #    'batting_CS', 'batting_BA', 'batting_OBP', 'batting_SLG', 'batting_TB',
+    #    'batting_SF', 'batting_SH', 'AU_POINTS', 'pitching_W', 'pitching_L',
+    #    'pitching_ERA', 'pitching_SHO', 'pitching_CG', 'pitching_SV',
+    #    'pitching_IP_str', 'pitching_IP', 'pitching_H', 'pitching_R',
+    #    'pitching_ER', 'pitching_HR', 'pitching_BB', 'pitching_SO',
+    #    'pitching_HBP', 'pitching_WP', 'pitching_WHIP', 'pitching_H9',
+    #    'pitching_HR9', 'pitching_BB9', 'pitching_SO9', 'pitching_SO/BB',
+    #    'pitching_RA9', 'pitching_PI', 'pitching_PI_balls',
+    #    'pitching_PI_strikes', 'fielding_position', 'fielding_IP_str',
+    #    'fielding_IP', 'fielding_PO', 'fielding_A', 'fielding_E', 'fielding_DP',
+    #    'fielding_FLD%', 'fielding_CS', 'fielding_CS%', 'fielding_TC',
+    #    'fielding_CH', 'fielding_RF/9']
+
+    finished_df = game_stats_df.groupby(['sport', 'api_version', 'season', 'seasonId','teamId'],as_index=False)[[
+        'G', 'AU_POINTS',
+        'batting_PA', 'batting_AB', 'batting_R',
+        'batting_H', 'batting_2B', 'batting_3B', 'batting_HR', 'batting_RBI',
+        'batting_BB', 'batting_HBP', 'batting_K', 'batting_SB', 'batting_SBA',
+        'batting_CS', 'batting_TB', 'batting_SF', 'batting_SH', 
+        'pitching_W', 'pitching_L', 'pitching_SHO', 'pitching_CG', 'pitching_QS',
+        'pitching_SV', 'pitching_IP', 
+        'pitching_H', 'pitching_R', 'pitching_ER', 'pitching_HR', 'pitching_BB', 
+        'pitching_SO', 'pitching_HBP', 'pitching_WP', 'pitching_PI', 'pitching_PI_balls',
+        'pitching_PI_strikes',
+        'fielding_IP', 'fielding_PO', 'fielding_A', 'fielding_E', 'fielding_DP',
+        'fielding_CS', 'fielding_TC'
+    ]].sum()
+
+    finished_df[['G',  'AU_POINTS','batting_PA', 'batting_AB', 'batting_R','batting_H', 'batting_2B', 'batting_3B', 'batting_HR', 'batting_RBI','batting_BB', 'batting_HBP', 'batting_K', 'batting_SB', 'batting_SBA','batting_CS', 'batting_TB', 'batting_SF', 'batting_SH', 'pitching_W', 'pitching_L', 'pitching_SHO', 'pitching_CG', 'pitching_QS','pitching_SV', 'pitching_H', 'pitching_R', 'pitching_ER', 'pitching_HR', 'pitching_BB', 'pitching_SO', 'pitching_HBP', 'pitching_WP', 'pitching_PI', 'pitching_PI_balls','pitching_PI_strikes','fielding_IP', 'fielding_PO', 'fielding_A', 'fielding_E', 'fielding_DP','fielding_CS', 'fielding_TC']] = finished_df[['G', 'AU_POINTS','batting_PA', 'batting_AB', 'batting_R','batting_H', 'batting_2B', 'batting_3B', 'batting_HR', 'batting_RBI','batting_BB', 'batting_HBP', 'batting_K', 'batting_SB', 'batting_SBA','batting_CS', 'batting_TB', 'batting_SF', 'batting_SH', 'pitching_W', 'pitching_L', 'pitching_SHO', 'pitching_CG', 'pitching_QS','pitching_SV','pitching_H', 'pitching_R', 'pitching_ER', 'pitching_HR', 'pitching_BB', 'pitching_SO', 'pitching_HBP', 'pitching_WP', 'pitching_PI', 'pitching_PI_balls','pitching_PI_strikes','fielding_IP', 'fielding_PO', 'fielding_A', 'fielding_E', 'fielding_DP','fielding_CS', 'fielding_TC']].astype('int')
+    finished_df['pitching_IP'] = finished_df['pitching_IP'].astype('float')
+    ## Batting
+    finished_df.loc[finished_df['batting_AB'] >= 1, 'batting_BA'] = finished_df['batting_H'] / finished_df['batting_AB']
+    finished_df['batting_BA'] = finished_df['batting_BA'].round(3)
+
+    finished_df.loc[finished_df['batting_AB'] > 0,'batting_OBP'] = (finished_df['batting_H'] + finished_df['batting_BB'] + finished_df['batting_HBP']) / (finished_df['batting_AB'] + finished_df['batting_BB'] + finished_df['batting_HBP'] + finished_df['batting_SF'])
+    finished_df['batting_OBP'] = finished_df['batting_OBP'].round(3)
+
+    finished_df.loc[finished_df['batting_AB'] > 0,'batting_SLG'] = finished_df['batting_TB'] / finished_df['batting_AB']
+    finished_df['batting_SLG'] = finished_df['batting_SLG'].round(3)
+
+    finished_df.loc[finished_df['batting_AB'] > 0,'batting_OPS'] = finished_df['batting_OBP'] + finished_df['batting_SLG']
+    finished_df['batting_OPS'] = finished_df['batting_OPS'].round(3)
+
+    finished_df['batting_OPS+'] = None
+
+    finished_df.loc[finished_df['batting_AB'] > 0,'batting_SecA'] = (finished_df['batting_BB'] + (finished_df['batting_TB'] - finished_df['batting_H']) + (finished_df['batting_SB'] - finished_df['batting_CS'])) / finished_df['batting_AB']
+    finished_df['batting_SecA'] = finished_df['batting_SecA'].round(3)
+
+    finished_df.loc[finished_df['batting_PA'] > 0,'batting_BB%'] = finished_df['batting_BB'] / finished_df['batting_PA']
+    finished_df['batting_BB%'] = finished_df['batting_BB%'].round(3)
+
+    finished_df.loc[finished_df['batting_PA'] > 0,'batting_K%'] = finished_df['batting_K'] / finished_df['batting_PA']
+    finished_df['batting_K%'] = finished_df['batting_K%'].round(3)
+    
+    finished_df.loc[finished_df['batting_AB'] > 0,'batting_ISO'] = (finished_df['batting_TB'] - finished_df['batting_H']) / finished_df['batting_AB']
+    finished_df['batting_ISO'] = finished_df['batting_ISO'].round(3)
+    
+    finished_df.loc[finished_df['batting_AB'] > 0,'batting_BABIP'] = (finished_df['batting_H'] + finished_df['batting_HR']) / (finished_df['batting_AB'] - finished_df['batting_K'] - finished_df['batting_HR'] + finished_df['batting_SF'])
+    finished_df['batting_BABIP'] = finished_df['batting_BABIP'].round(3)
+
+    finished_df.loc[(finished_df['batting_SB'] > 0) | (finished_df['batting_HR'] > 0),'batting_PSN'] = (2 * finished_df['batting_HR'] * finished_df['batting_SB']) / (finished_df['batting_HR'] * finished_df['batting_SB'])
+    finished_df['batting_PSN'] = finished_df['batting_PSN'].round(3)
+
+    ## Pitching
+    finished_df['pitching_ERA'] = 9 * (finished_df['pitching_ER'] / finished_df['pitching_IP'])
+    finished_df['pitching_ERA'] = finished_df['pitching_ERA'].round(3)
+
+    finished_df['pitching_ERA+'] = None
+    finished_df['pitching_FIP'] = None
+    finished_df['pitching_FIP-'] = None
+    finished_df['pitching_WHIP'] = (finished_df['pitching_BB'] + finished_df['pitching_H']) / finished_df['pitching_IP']
+    finished_df['pitching_WHIP'] = finished_df['pitching_WHIP'].round(3)
+
+    finished_df['pitching_H9'] = (9 * finished_df['pitching_H']) / finished_df['pitching_IP']
+    finished_df['pitching_H9'] = finished_df['pitching_H9'].round(3)
+
+    finished_df['pitching_HR9'] = (9 * finished_df['pitching_HR']) / finished_df['pitching_IP']
+    finished_df['pitching_HR9'] = finished_df['pitching_HR9'].round(3)
+
+    finished_df['pitching_BB9'] = (9 * finished_df['pitching_BB']) / finished_df['pitching_IP']
+    finished_df['pitching_BB9'] = finished_df['pitching_BB9'].round(3)
+
+    finished_df['pitching_SO9'] = (9 * finished_df['pitching_SO']) / finished_df['pitching_IP']
+    finished_df['pitching_SO9'] = finished_df['pitching_SO9'].round(3)
+
+    finished_df['pitching_SO/BB'] = finished_df['pitching_SO'] / finished_df['pitching_BB']
+    finished_df['pitching_SO/BB'] = finished_df['pitching_SO/BB'].round(3)
+
+    finished_df['pitching_RA9'] = 9 * (finished_df['pitching_R'] / finished_df['pitching_IP'])
+    finished_df['pitching_RA9'] = finished_df['pitching_RA9'].round(3)
+    
+    ## Fielding
+    finished_df['fielding_FLD%'] = (finished_df['fielding_PO'] + finished_df['fielding_A']) / (finished_df['fielding_PO'] + finished_df['fielding_A'] + finished_df['fielding_E'])
+    finished_df['fielding_FLD%'] = finished_df['fielding_FLD%'].round(3)
+
+    finished_df['fielding_CH'] = finished_df['fielding_PO'] + finished_df['fielding_A'] + finished_df['fielding_E']
+    finished_df['fielding_CH'] = finished_df['fielding_CH'].round(3)
+
+    #finished_df['fielding_CS%'] = 0
+    #finished_df['fielding_CS%'] = finished_df['fielding_CS%'].round(3)
+
+    finished_df['fielding_RF/9'] = (9 * (finished_df['fielding_PO'] + finished_df['fielding_A'])) / finished_df['fielding_IP']
+    finished_df['fielding_RF/9'] = finished_df['fielding_RF/9'].round(3)
+
+    # finished_df.to_csv('test.csv')
+    # print(finished_df)
+    # print(game_stats_df.columns)
+    return finished_df